test: history_service -> add unit tests for service

+++ history_service/internal/revision.go
@@ -4,7 +4,7 @@ import "time"
 
 type Revision struct {
 	RevisionNumber string    `json:"revisionNumber" bson:"revisionNumber"`
-	ProductID      int64     `json:"productID" bson:"productID"`
+	ProductID      uint      `json:"productID" bson:"productID"`
 	ChangedAttr    []string  `json:"changedAttr" bson:"changedAttr"`
 	PrevValue      *Product  `json:"prevValue" bson:"prevValue"`
 	NewValue       *Product  `json:"newValue" bson:"newValue"`
